Tickets #347, #349, #387, #393, #394. Moving r4131 to production.
git-svn-id: file:///home/jaliste/hcvol/hcvol/branches/production@4132 4f49b16c-3c57-40ef-8e22-7452baa6c33a

+++ build/gallery/gallery.model.php
@@ -50,18 +50,26 @@ VALUES
         return true;
     }
     
-    // delete own uploaded pictures as logged in user
+    // delete own uploaded pictures as logged in user or user with gallery rights
     public function deleteOneProcess($image)
     {
         if (!$User = APP_User::login())
             return false;
-        $filename = $image->file;
-        $userDir = new PDataDir('gallery/user'.$User->getId());
-        $userDir->delFile($filename);
-        $userDir->delFile('thumb'.$filename);
-        $userDir->delFile('thumb2'.$filename);
-        $this->dao->exec('DELETE FROM `gallery_items` WHERE `id` = '.$image->id);
-        return;
+        $R = MOD_right::get();
+        $GalleryRight = $R->hasRight('Gallery');
+        if (($User->getId() == $this->imageOwner($image->id)) || ($GalleryRight > 1)) {
+            // Log the deletion to prevent admin abuse
+            MOD_log::get()->write("Deleting a gallery item #".$image->id." filename: ".$image->file." belonging to user: ".$image->user_id_foreign, "Gallery");
+            // Start the deletion process
+            $filename = $image->file;
+            $userDir = new PDataDir('gallery/user'.$image->user_id_foreign);
+            $userDir->delFile($filename);
+            $userDir->delFile('thumb'.$filename);
+            $userDir->delFile('thumb2'.$filename);
+            $this->dao->exec('DELETE FROM `gallery_items` WHERE `id` = '.$image->id);
+            $this->deleteComments($image->id);
+            return true;
+        } else return false;
     }
     
     public function editProcess()
@@ -71,13 +79,94 @@ VALUES
             if (!$User = APP_User::login())
                 return false;
             $vars = &PPostHandler::getVars($callbackId);
-            $this->dao->exec("UPDATE `gallery_items` SET `title` = '".$vars['t']."' WHERE `id`= ".$vars['id']);
+            $this->dao->exec("UPDATE `gallery_items` SET `title` = '".$vars['t']."' , `description` = '".$vars['txt']."' WHERE `id`= ".$vars['id']);
             PPostHandler::clearVars($callbackId);
         	return false;
         } else {
         	PPostHandler::setCallback($callbackId, __CLASS__, __FUNCTION__);
             return $callbackId;
         }
+    }
+
+    /**
+     * Processing creation of a comment
+     *
+     * This is a POST callback function.
+     *
+     * Sets following errors in POST vars:
+     * title        - invalid(empty) title.
+     * textlen      - too short or long text.
+     * inserror     - db error while inserting.
+     */
+    public function commentProcess($image = false) {
+    	$callbackId = PFunctions::hex2base64(sha1(__METHOD__));
+        if (PPostHandler::isHandling()) {
+            if (!$User = APP_User::login())
+                return false;
+            $vars =& PPostHandler::getVars();
+            $request = PRequest::get()->request;
+            if (!$image)
+                $image = (int)$request[3];
+
+            // validate
+            if (!isset($vars['ctxt']) || strlen($vars['ctxt']) == 0 || strlen($vars['ctxt']) > 5000) {
+                $vars['errors'] = array('textlen');
+                return false;
+            }
+
+            $commentId = $this->dao->nextId('gallery_comments');
+            $query = '
+INSERT INTO `gallery_comments`
+SET
+    `id`='.$commentId.',
+    `gallery_items_id_foreign`='.$image.',
+    `user_id_foreign`='.$User->getId().',
+    `title`=\''.(isset($vars['ctit'])?$this->dao->escape($vars['ctit']):'').'\',
+    `text`=\''.$this->dao->escape($vars['ctxt']).'\',
+    `created`=NOW()';
+            $s = $this->dao->query($query);
+            if (!$s) {
+                $vars['errors'] = array('inserror');
+                return false;
+            }
+            PPostHandler::clearVars();
+            return PVars::getObj('env')->baseuri.implode('/', $request).'#c'.$commentId;
+        } else {
+        	PPostHandler::setCallback($callbackId, __CLASS__, __FUNCTION__);
+            return $callbackId;
+        }
+    }
+
+    
+    /*
+         Adding comments
+        */
+    public function getComments($image) {
+    	$query = '
+SELECT
+    c.`id` AS `comment_id`,
+    c.`user_id_foreign` AS `user_id`,
+    u.`handle` AS `user_handle`,
+    UNIX_TIMESTAMP(c.`created`) AS `unix_created`,
+    c.`title`,
+    c.`created`,
+    c.`text`
+FROM `gallery_comments` c
+LEFT JOIN `user` u ON c.`user_id_foreign`=u.`id`
+WHERE c.`gallery_items_id_foreign` = '.(int)$image.'
+        ';
+        $s = $this->dao->query($query);
+        if ($s->numRows() == 0)
+            return false;
+        return $s;
+    }
+
+    public function deleteComments($image) {
+    	$query = '
+DELETE FROM `gallery_comments`
+WHERE `gallery_items_id_foreign` = '.(int)$image.'
+        ';
+        return $this->dao->exec($query);
     }
 
     public function getLatestItems($userId = false)
@@ -127,6 +216,7 @@ SELECT
     i.`width`, 
     i.`height`, 
     i.`title`,
+    i.`description`,
     i.`created`
 FROM `gallery_items` AS i
 LEFT JOIN `user` AS `u` ON
@@ -233,8 +323,8 @@ VALUES
                 $vars['error'] = 'Gallery_UploadError';
                 return false;
             }
-            if ($size[0] > 400)
-                $img->createThumb($userDir->dirName(), 'thumb2', 350);
+            if ($size[0] > 550)
+                $img->createThumb($userDir->dirName(), 'thumb2', 500);
             $itemId = $this->dao->nextId('gallery_items');
             $orig = $_FILES['gallery-file']['name'];
             $mimetype = image_type_to_mime_type($type);
